changes to the form to upload papers
